requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.12.0
django-jalali-date==1.1.0
django-ckeditor==6.7.1